+++ compiler/src/emit/native.rs
@@ -301,7 +301,7 @@ pub(crate) fn emit_natives(
             instructions.emit_code(Opcode::ConvertByteToInt);
         }
         52 => {
-            // Vec[A]::shift() -> Option[A]
+            // Vec[A]::pop_head() -> Option[A]
             instructions.emit_invoke(cp.insert_string(VEC_POP_HEAD));
         }
         id => todo!("Native function with id {id}"),
