+++ lib/std/memory.msh
@@ -5,5 +5,5 @@ fun gc();
 /// return true if the thread's operands stack is empty
 fun empty_operands() -> Bool;
 
-/// return a vector containing this process' arguments
+/// Returns a vector containing the arguments of this process.
 fun program_arguments() -> Vec[String];
